Correct alumni (see long commit message)
Following an impromptu discussion with some faculty and Irene, for now
we are going to list only students who have formally graduated with
some graduate degree. This excludes the following:
- ABD students
- Transfers midway through a Penn degree (unless a Masters is awarded)
- Other students who left the program without a degree

Part of the motiviation for this is that we do not have their current
information (e.g. website, contact email) and we don't have a clear
picture of what we would list on the website. (Some students we have
information for, but I think a consistent policy is better.)