Cleanup people folder/ (#27)
* Edit template: Do not show broken links if website data is missing

Previously when a student's "website" field is missing, we were
treating the field like an empty string. Browsers were filling this in
with an href that pointed back to the PL Club website. Now we do not
try to show that data when it is not available.

**NOTE**: We did not make similar edits to faculty or postdocs because
presently nobody in those categories does not have a website. In the
future this may not hold.

* Delete obsolete people/orig/ folder

* Remove "Add description here." from people files

We are not currently using the body of people files and don't want to
confuse anyone who wants to make edits to them.